requirements.txt
python-telegram-bot==20.7
openai==1.30.1
numpy==1.26.4
python-dotenv==1.0.1
